Drop the NaNs from the groundtruth parquet before writing
This is the cause of the differences between the inference results reported
by seb and the results returned by this file. Previously the `dropna` call
was happening after the groundtruth parquet had been reduced to two columns
which never contain any NaNs - It seems that the NaN columns in the full
groundtruth parquet were not used in fine-tuning the model?